feat: Enable custom styling of OTP Form (ACQ-1673) (#47)
Enable custom styling of the new displayed OTP Form within the `iframe` through the `theme` option. You can target any of the elements with the `theme` option:
```js
{
  theme: {
    formOTP: { /* ... */ } ,
    formOTPInput: { /* ... */ },
    formOTPSubmit: { /* ... */ },
}
```
We also added a **default dark theme** for the OTP Form and fixed a couple of minor styling issues on both Chrome and Safari.